myproject/requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distlib==0.3.9
Django==5.1.5
djangorestframework==3.15.2
dnspython==2.7.0
filelock==3.17.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
platformdirs==4.3.6
pymongo==4.10.1
requests==2.32.3
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
virtualenv==20.29.1
Werkzeug==3.1.3
